change: rename imagerules integration tests
