+++ integration/client/imagerules/util.go
@@ -1,4 +1,4 @@
-package signatures
+package imagerules
 
 import (
 	"context"
